requirements.txt
Django==1.11.7
Landscape-Client==14.12
PAM==0.4.2
Twisted-Core==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
beautifulsoup4==4.2.1
chardet==2.0.1
colorama==0.2.5
configobj==4.7.2
django-graphos==0.3.41
gyp==0.1
html5lib==0.999
nodeenv==1.2.0
pbr==3.1.1
psycopg2==2.7.3.2
pyOpenSSL==0.13
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
pytz==2017.3
requests==2.2.1
six==1.11.0
ssh-import-id==3.21
stevedore==1.27.1
urllib3==1.7.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
